+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.0
 
 - Make `TileMap.where` return an `Iterable`.
+- Make `TileMap.tilesInRange` return an `Iterable`.
 
 ## 0.0.1